Move fine-tuning CL pipeline
